# apps/ai-agents/openapi_resp.json
{"openapi":"3.1.0","info":{"title":"MUNify AI Agents API","description":"API for triggering LangGraph AI agents to generate MUN documents","version":"2.0.0"},"paths":{"/api/v1/generate":{"post":{"summary":"Generate Document","operationId":"generate_document_api_v1_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/convert/html-to-latex":{"post":{"summary":"Convert Html To Latex Endpoint","description":"Convierte HTML del editor Tiptap a LaTeX para persistencia.","operationId":"convert_html_to_latex_endpoint_api_v1_convert_html_to_latex_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConvertRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConvertResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/compile-pdf":{"post":{"summary":"Compile Pdf","description":"Compila LaTeX a PDF usando pdflatex y devuelve el archivo.","operationId":"compile_pdf_api_v1_compile_pdf_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompileRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/compile-pdf-from-html":{"post":{"summary":"Compile Pdf From Html","description":"Convierte HTML a LaTeX, lo envuelve en una plantilla y compila a PDF.","operationId":"compile_pdf_from_html_api_v1_compile_pdf_from_html_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompileHtmlRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ingest-context":{"post":{"summary":"Ingest Context","description":"Crea un notebook y sube archivos PDF para contexto.","operationId":"ingest_context_api_v1_ingest_context_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CompileHtmlRequest":{"properties":{"html":{"type":"string","title":"Html"},"title":{"type":"string","title":"Title"},"topic":{"type":"string","title":"Topic"},"country":{"type":"string","title":"Country"},"committee":{"type":"string","title":"Committee"}},"type":"object","required":["html","title","topic","country","committee"],"title":"CompileHtmlRequest"},"CompileRequest":{"properties":{"latex":{"type":"string","title":"Latex"}},"type":"object","required":["latex"],"title":"CompileRequest"},"ConvertRequest":{"properties":{"content":{"type":"string","title":"Content"}},"type":"object","required":["content"],"title":"ConvertRequest"},"ConvertResponse":{"properties":{"latex":{"type":"string","title":"Latex"}},"type":"object","required":["latex"],"title":"ConvertResponse"},"GenerateRequest":{"properties":{"topic":{"type":"string","title":"Topic"},"country":{"type":"string","title":"Country"},"committee":{"type":"string","title":"Committee"},"documentType":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Documenttype","default":"POSITION_PAPER"},"notebookId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notebookid"}},"type":"object","required":["topic","country","committee"],"title":"GenerateRequest"},"GenerateResponse":{"properties":{"draft":{"type":"string","title":"Draft"},"draft_html":{"type":"string","title":"Draft Html"},"strategy_guide":{"type":"string","title":"Strategy Guide"},"errors":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Errors"},"research_data":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Research Data"}},"type":"object","required":["draft","draft_html","strategy_guide"],"title":"GenerateResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IngestRequest":{"properties":{"notebookTitle":{"type":"string","title":"Notebooktitle"},"files":{"items":{"type":"string"},"type":"array","title":"Files"}},"type":"object","required":["notebookTitle","files"],"title":"IngestRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}